+++ src/ui/components/tests/layer-avatar-test.js
@@ -132,6 +132,38 @@ describe('layer-avatar', function() {
     expect(el.classList.contains('layer-avatar-cluster')).toBe(true);
   });
 
+  it("Should set layer-empty-avatar on nameless identity", function() {
+    el.users = [
+      new Layer.Core.Identity({userId: 'AA', id: 'layer:///identities/AA'})
+    ];
+    expect(el.childNodes[0].classList.contains('layer-empty-avatar')).toBe(true);
+
+    el.users = [
+      new Layer.Core.Identity({userId: 'AA', id: 'layer:///identities/AA'}),
+      new Layer.Core.Identity({userId: 'BB', id: 'layer:///identities/BB', displayName: "BB"})
+    ];
+    expect(el.childNodes[0].classList.contains('layer-empty-avatar')).toBe(false);
+    expect(el.childNodes[1].classList.contains('layer-empty-avatar')).toBe(true);
+  });
+
+  it("Should set layer-empty-group-avatar on nameless identities", function() {
+    el.users = [
+      new Layer.Core.Identity({userId: 'AA', id: 'layer:///identities/AA'}),
+      new Layer.Core.Identity({userId: 'BB', id: 'layer:///identities/BB'})
+    ];
+    expect(el.childNodes[0].classList.contains('layer-empty-group-avatar')).toBe(true);
+    expect(el.childNodes.length).toEqual(1);
+
+    el.users = [
+      new Layer.Core.Identity({userId: 'AA', id: 'layer:///identities/AA'}),
+      new Layer.Core.Identity({userId: 'BB', id: 'layer:///identities/BB'}),
+      new Layer.Core.Identity({userId: 'CC', id: 'layer:///identities/CC', displayName: "CC"})
+    ];
+    expect(el.childNodes[0].classList.contains('layer-empty-group-avatar')).toBe(false);
+    expect(el.childNodes[0].classList.contains('layer-empty-avatar')).toBe(false);
+    expect(el.childNodes[1].classList.contains('layer-empty-group-avatar')).toBe(true);
+  });
+
   it("Should have a layer-presence widget with its user set", function() {
     el.users = [client.user];
     expect(el.nodes.presence.tagName).toEqual('LAYER-PRESENCE');