+++ CHANGELOG.md
@@ -12,7 +12,7 @@ Additional Changes:
   from the server.
 * WEB-1421: Client now throws errors if a `challenge` event is not registered.  All apps should register this event even if only to handle reauthentication
 * WEB-1198: Filters out invalid Identities from `createConversation` and `addParticipants`
-
+* WEB-1619: Adds Anonymous avatar and group avatar graphics
 
 ## 1.0.0-pre2.8